requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.0
python-dotenv==1.0.0
marshmallow==3.20.1
werkzeug==3.0.1
pytest==7.4.3
requests==2.31.0
pandas==2.1.3
reportlab==4.0.7
gunicorn==21.2.0    